Minor updates. Making sure the code is ready for a PR.
